requirements.txt
attrs==25.1.0
beautifulsoup4==4.13.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
py4j==0.10.9.9
pyparsing==3.2.1
PySocks==1.7.1
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-whois==0.9.5
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
selenium==4.28.1
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sympy==1.13.1
tabulate==0.9.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.91
ultralytics-thop==2.0.14
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
